Readme edits and naming convention suggestion add
Added a few details to the instructions for the places I tripped over a little. Also added naming convention suggestions for the metadata. I think it may be worth systemizing a naming scheme, it got out of hand for me immediately. Especially as end users will be interacting with this tool I think a standard would help prevent confusion and naming chaos.